+++ include/widget_value.hpp
@@ -6,7 +6,7 @@
 
 class WidgetValue : public Widget {
 public:
-  WidgetValue(LuaSaveParser *saveParser, u64 minValue, u64 maxValue);
+  WidgetValue(LuaSaveParser *saveParser, s32 minValue, s32 maxValue);
   ~WidgetValue();
 
   void draw(Gui *gui, u16 x, u16 y);
@@ -15,5 +15,5 @@ public:
   void onTouch(touchPosition &touch);
 
 private:
-  u64 m_minValue, m_maxValue;
+  s32 m_minValue, m_maxValue;
 };
